feat : getArchiveId 메서드 수정
